+++ config/binding-http.conf/src/main/java/io/aklivity/zilla/config/binding/http/HttpOptionsConfig.java
@@ -14,17 +14,12 @@
  */
 package io.aklivity.zilla.config.binding.http;
 
-import static java.util.Collections.emptyList;
-
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.SortedSet;
 import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-import io.aklivity.zilla.config.engine.ModelConfig;
+import io.aklivity.zilla.config.engine.NamedConfig;
 import io.aklivity.zilla.config.engine.OptionsConfig;
 
 public final class HttpOptionsConfig extends OptionsConfig
@@ -51,48 +46,14 @@ public final class HttpOptionsConfig extends OptionsConfig
         Map<String, String> overrides,
         HttpAccessControlConfig access,
         HttpAuthorizationConfig authorization,
-        List<HttpRequestConfig> requests)
+        List<HttpRequestConfig> requests,
+        List<NamedConfig> refs)
     {
-        super(resolveModels(requests), List.of());
+        super(null, refs);
         this.versions = versions;
         this.overrides = overrides;
         this.access = access;
         this.authorization = authorization;
         this.requests = requests;
-    }
-
-    private static List<ModelConfig> resolveModels(
-        List<HttpRequestConfig> requests)
-    {
-        return requests != null && !requests.isEmpty()
-            ? requests.stream()
-            .flatMap(request -> Stream.concat(
-                Stream.of(request.content),
-                Stream.concat(
-                    request.headers != null
-                        ? request.headers.stream().flatMap(header -> Stream.of(header != null ? header.model : null))
-                        : Stream.empty(),
-                    Stream.concat(
-                        request.pathParams != null
-                            ? request.pathParams.stream().flatMap(param -> Stream.of(param != null ? param.model : null))
-                            : Stream.empty(),
-                        Stream.concat(
-                            request.queryParams != null
-                                ? request.queryParams.stream().flatMap(param -> Stream.of(param != null ? param.model : null))
-                                : Stream.empty(),
-                            Stream.concat(request.responses != null
-                                ? request.responses.stream().flatMap(param -> Stream.of(param != null
-                                ? param.content
-                                : null))
-                                : Stream.empty(), request.responses != null
-                                ? request.responses.stream()
-                                .flatMap(response -> response.headers != null
-                                    ? response.headers.stream()
-                                    .flatMap(param -> Stream.of(param != null ? param.model : null))
-                                    : Stream.empty())
-                                : Stream.empty())
-                        )))).filter(Objects::nonNull))
-            .collect(Collectors.toList())
-            : emptyList();
     }
 }