+++ config/binding-kafka.conf/src/main/java/io/aklivity/zilla/config/binding/kafka/KafkaTopicConfigBuilder.java
@@ -14,10 +14,13 @@
  */
 package io.aklivity.zilla.config.binding.kafka;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.function.Function;
 
 import io.aklivity.zilla.config.engine.ConfigBuilder;
 import io.aklivity.zilla.config.engine.ModelConfig;
+import io.aklivity.zilla.config.engine.NamedConfig;
 
 public final class KafkaTopicConfigBuilder<T> extends ConfigBuilder<T, KafkaTopicConfigBuilder<T>>
 {
@@ -105,6 +108,15 @@ public final class KafkaTopicConfigBuilder<T> extends ConfigBuilder<T, KafkaTopi
     @Override
     public T build()
     {
-        return mapper.apply(new KafkaTopicConfig(name, defaultOffset, deltaType, key, value, transforms));
+        List<NamedConfig> refs = new ArrayList<>();
+        if (key != null)
+        {
+            refs.addAll(key.refs());
+        }
+        if (value != null)
+        {
+            refs.addAll(value.refs());
+        }
+        return mapper.apply(new KafkaTopicConfig(name, defaultOffset, deltaType, key, value, transforms, refs));
     }
 }