+++ config/binding-mcp-kafka-connect.conf/src/main/java/io/aklivity/zilla/config/binding/mcp/kafka/connect/McpKafkaConnectOptionsConfig.java
@@ -36,6 +36,7 @@ public final class McpKafkaConnectOptionsConfig extends OptionsConfig
     McpKafkaConnectOptionsConfig(
         String server)
     {
+        super(null, null);
         this.server = server;
     }
 }